rename starkware_patch to starkware
