--- src/commands/test/starkware_patch/__init__.py
@@ -1,6 +1,0 @@
-from src.commands.test.starkware_patch.cheatable_syscall_handler import (
-    CheatableHintsWhitelist,
-    CheatableSysCallHandler,
-    CheatcodeException,
-)
-from src.commands.test.starkware_patch.forkable_starknet import ForkableStarknet
